move InternalClientOptions and NodeInitializationParameters to types modules
