+++ modules/client/src/node.ts
@@ -12,9 +12,11 @@ import {
   GetChannelResponse,
   GetConfigResponse,
   IChannelProvider,
+  INodeApiClient,
   makeChecksumOrEthAddress,
   NodeInitializationParameters,
   PaymentProfile,
+  PendingAsyncTransfer,
   RequestCollateralResponse,
   ResolveLinkedTransferResponse,
   SupportedApplication,
@@ -26,35 +28,6 @@ import { invalidXpub } from "./validation";
 // Include our access token when interacting with these subjects
 const guardedSubjects = ["channel", "lock", "transfer"];
 const sendFailed = "Failed to send message";
-
-export interface INodeApiClient {
-  acquireLock(lockName: string, callback: (...args: any[]) => any, timeout: number): Promise<any>;
-  addPaymentProfile(profile: PaymentProfile): Promise<PaymentProfile>;
-  appRegistry(appDetails?: {
-    name: SupportedApplication;
-    network: SupportedNetwork;
-  }): Promise<AppRegistry>;
-  config(): Promise<GetConfigResponse>;
-  createChannel(): Promise<CreateChannelResponse>;
-  getChannel(): Promise<GetChannelResponse>;
-  getLatestSwapRate(from: string, to: string): Promise<string>;
-  getPaymentProfile(assetId?: string): Promise<PaymentProfile>;
-  getTransferHistory(publicIdentifier: string): Promise<Transfer[]>;
-  requestCollateral(assetId: string): Promise<RequestCollateralResponse | void>;
-  withdraw(tx: CFCoreTypes.MinimalTransaction): Promise<TransactionResponse>;
-  fetchLinkedTransfer(paymentId: string): Promise<any>;
-  resolveLinkedTransfer(
-    paymentId: string,
-    linkedHash: string,
-    meta: object,
-  ): Promise<ResolveLinkedTransferResponse>;
-  recipientOnline(recipientPublicIdentifier: string): Promise<boolean>;
-  restoreState(publicIdentifier: string): Promise<any>;
-  subscribeToSwapRates(from: string, to: string, callback: any): void;
-  unsubscribeFromSwapRates(from: string, to: string): void;
-  // TODO: fix types
-  verifyAppSequenceNumber(appSequenceNumber: number): Promise<ChannelAppSequences>;
-}
 
 // NOTE: swap rates are given as a decimal string describing:
 // Given 1 unit of `from`, how many units `to` are recieved.
@@ -144,15 +117,7 @@ export class NodeApiClient implements INodeApiClient {
     return await this.send(`channel.get.${this.userPublicIdentifier}`);
   }
 
-  public async getPendingAsyncTransfers(): Promise<
-    {
-      assetId: string;
-      amount: string;
-      encryptedPreImage: string;
-      linkedHash: string;
-      paymentId: string;
-    }[]
-  > {
+  public async getPendingAsyncTransfers(): Promise<PendingAsyncTransfer[]> {
     return (await this.send(`transfer.get-pending.${this.userPublicIdentifier}`)) || [];
   }
