+++ modules/types/src/index.ts
@@ -6,4 +6,5 @@ export * from "./channelProvider";
 export * from "./client";
 export * from "./convert";
 export * from "./inputs";
+export * from "./messaging";
 export * from "./node";
