+++ modules/types/src/node.ts
@@ -1,15 +1,12 @@
-import { BigNumber, Network } from "ethers/utils";
+import { TransactionResponse } from "ethers/providers";
+import { BigNumber, Network, Transaction } from "ethers/utils";
 
-import { SupportedApplication } from "./app";
+import { AppRegistry, SupportedApplication, SupportedNetwork } from "./app";
 import { CFCoreTypes, NetworkContext } from "./cf";
-import { CFCoreChannel } from "./channel";
-
-export interface MessagingConfig {
-  clusterId?: string;
-  messagingUrl: string | string[];
-  token?: string;
-  logLevel: number;
-}
+import { CFCoreChannel, ChannelAppSequences, PaymentProfile } from "./channel";
+import { IChannelProvider } from "./channelProvider";
+import { ResolveLinkedTransferResponse } from "./inputs";
+import { IMessagingService, MessagingConfig } from "./messaging";
 
 ////////////////////////////////////
 ///////// NODE RESPONSE TYPES
@@ -65,3 +62,63 @@ export type CreateChannelResponse = {
 
 // TODO: why was this changed?
 export type RequestCollateralResponse = CFCoreTypes.DepositResult | undefined;
+
+////////////////////////////////////
+///////// NODE API CLIENT
+
+export interface PendingAsyncTransfer {
+  assetId: string;
+  amount: string;
+  encryptedPreImage: string;
+  linkedHash: string;
+  paymentId: string;
+}
+
+export interface NodeInitializationParameters {
+  messaging: IMessagingService;
+  logLevel?: number;
+  userPublicIdentifier?: string;
+  nodePublicIdentifier?: string;
+  channelProvider?: IChannelProvider;
+}
+
+export interface INodeApiClient {
+  channelProvider: IChannelProvider | undefined;
+  userPublicIdentifier: string | undefined;
+  nodePublicIdentifier: string | undefined;
+
+  acquireLock(lockName: string, callback: (...args: any[]) => any, timeout: number): Promise<any>;
+  addPaymentProfile(profile: PaymentProfile): Promise<PaymentProfile>;
+  appRegistry(appDetails?: {
+    name: SupportedApplication;
+    network: SupportedNetwork;
+  }): Promise<AppRegistry>;
+  config(): Promise<GetConfigResponse>;
+  createChannel(): Promise<CreateChannelResponse>;
+  clientCheckIn(): Promise<void>;
+  getChannel(): Promise<GetChannelResponse>;
+  getLatestSwapRate(from: string, to: string): Promise<string>;
+  getPaymentProfile(assetId?: string): Promise<PaymentProfile>;
+  getPendingAsyncTransfers(): Promise<PendingAsyncTransfer[]>;
+  getTransferHistory(publicIdentifier?: string): Promise<Transfer[]>;
+  getLatestWithdrawal(): Promise<Transaction>;
+  requestCollateral(assetId: string): Promise<RequestCollateralResponse | void>;
+  withdraw(tx: CFCoreTypes.MinimalTransaction): Promise<TransactionResponse>;
+  fetchLinkedTransfer(paymentId: string): Promise<any>;
+  resolveLinkedTransfer(
+    paymentId: string,
+    linkedHash: string,
+    meta: object,
+  ): Promise<ResolveLinkedTransferResponse>;
+  recipientOnline(recipientPublicIdentifier: string): Promise<boolean>;
+  restoreState(publicIdentifier: string): Promise<any>;
+  subscribeToSwapRates(from: string, to: string, callback: any): void;
+  unsubscribeFromSwapRates(from: string, to: string): void;
+  // TODO: fix types
+  verifyAppSequenceNumber(appSequenceNumber: number): Promise<ChannelAppSequences>;
+  setRecipientAndEncryptedPreImageForLinkedTransfer(
+    recipient: string,
+    encryptedPreImage: string,
+    linkedHash: string,
+  ): Promise<any>;
+}
